Refactor API routing: centralize routers and reorganize v1 endpoints
- Move v1 route files (`roles`, `users`, `role_assignments`) into
  dedicated subdirectories with `route.py` and `__init__.py` for
  improved modularity
- Update import paths in `any_auth/api/v1/route.py` to reflect new
  structure
- Introduce `any_auth/api/route.py` to consolidate all API routers
  into a single entry point
- Refactor `build_app.py` to include the new centralized API router,
  simplifying route registration and reducing duplication

This refactor enhances maintainability, clarity, and scalability of
the API routing structure.
